refactor : itemsState와 itemDetailsState에서 중복되는 데이터 삭제
